Add model caching (closes #18, closes #23)
I am not perfectly happy with this solution.  The whole
DepthFirstExplorer.refine() needs a rewrite.  The logic is just a little
too fuzzy.  In the end, the code may turn out to be the same though.

The model caching does not avoid all revisits, but at least it
eliminates duplicate models (apart from duplicating the very first run's
inputs at most once).