Invoke delegates for static methods
The INVOKESPECIAL and INVOKEVIRTUAL methods still need to be changed to
make use of delegated methods.  For the moment, only INVOKESTATIC does
so.  However, the first two instructions were updated to use the new
method in the Symbolizer to create new variables.  This means that there
is a global variable count and less chance for naming conflicts.