Add examples of how delegates are used
For the moment, the Math delegate does not add its additional guard to
the path condition.

Firstly, GREEN needs an update to handle such guards which may contain
complex expressions.  This involves not simplifying the expression and
passing it to Z3 untouched.  In the long run, GREEN should also simplify
complex expressions, but it has never been used in this new context and
it does not currently implement this functionality.

Secondly, DEEPSEA needs to be adjusted to handle the conjunction of
additional guards separately and pass it to the Explorer as an
additional path condition.  The Explorer should pass the conjunction of
the regular old path condition and the additional path condition to the
solver (GREEN), and then it should negate only the regular old path
condition's conjuncts.  This is not a big change and the code is already
in place, but it needs to be "activated".