Add "thread" parameter and "getRealArrayLength" routine.
Delegated methods must now accept a "thread" parameter (which they can
ignore).  It is useful for array parameters (and anything else that
wants to inspect the memory space of the VM under test).  For example, a
new static Symbolizer.getRealArrayLength routine returns the length of
an array on the actual stack.