Modify delegateMethods to allow for real argument inspection.
Before, delegates are invoked when the corresponding INVOKE...
instruction executed.  However, at this point, the frame of the concrete
VM has not yet been set up, and it is not possible to inspect the
argument values.  We need these values to determine, e.g., the size of
arrays.

After, delegates are found and "remembered" in class fields, and only
invoked when the MethodEntry event is triggered.